# internal/assets/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tumble API",
    "description": "API for the Tumble content application. The v1 API provides a RESTful, versioned interface with consistent request/response formats. Legacy endpoints are deprecated and will be removed in a future release.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for protected endpoints"
      }
    },
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "validation_error", "internal_error"]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Field-level validation errors (for validation_error code)"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"],
        "example": {
          "error": {
            "code": "not_found",
            "message": "Link 42 does not exist"
          }
        }
      },
      "APIMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of items available"
          },
          "limit": {
            "type": "integer",
            "description": "Maximum items per page"
          },
          "offset": {
            "type": "integer",
            "description": "Number of items skipped"
          }
        },
        "required": ["total", "limit", "offset"]
      },
      "APILinkResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique link ID"
          },
          "url": {
            "type": "string",
            "description": "The target URL"
          },
          "title": {
            "type": "string",
            "description": "Title of the linked page"
          },
          "user": {
            "type": "string",
            "description": "Username who submitted the link"
          },
          "clicks": {
            "type": "integer",
            "description": "Number of verified clicks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the link was submitted"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with this link"
          }
        },
        "required": ["id", "url", "user", "created_at"],
        "example": {
          "id": 42,
          "url": "https://example.com/article",
          "title": "Example Article",
          "user": "alice",
          "clicks": 15,
          "created_at": "2026-01-15T10:30:00Z",
          "tags": ["golang", "tutorial"]
        }
      },
      "APILinkCreateResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APILinkResponse" },
          {
            "type": "object",
            "properties": {
              "is_duplicate": {
                "type": "boolean",
                "description": "Whether this URL was previously submitted"
              },
              "previous_submissions": {
                "type": "array",
                "description": "List of prior submissions if duplicate",
                "items": {
                  "$ref": "#/components/schemas/APIPreviousSubmission"
                }
              }
            }
          }
        ],
        "example": {
          "id": 456,
          "url": "https://example.com/article",
          "title": "Example Article",
          "user": "bob",
          "clicks": 0,
          "created_at": "2026-02-08T14:30:00Z",
          "is_duplicate": true,
          "previous_submissions": [
            {
              "id": 123,
              "user": "alice",
              "created_at": "2026-01-15T10:30:00Z",
              "title": "Example Article"
            }
          ]
        }
      },
      "APIPreviousSubmission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID of the previous submission"
          },
          "user": {
            "type": "string",
            "description": "Username who submitted previously"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the link was previously submitted"
          },
          "title": {
            "type": "string",
            "description": "Title of the previous submission"
          }
        },
        "required": ["id", "user", "created_at"]
      },
      "APILinksListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APILinkResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        },
        "required": ["data", "meta"]
      },
...
      "LinkCreateRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL to submit"
          },
          "user": {
            "type": "string",
            "description": "Username of the submitter"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags to add to the link (lowercased, no spaces)"
          }
        },
        "required": ["url", "user"],
        "example": {
          "url": "https://example.com/article",
          "user": "alice",
          "tags": ["golang", "tutorial"]
        }
      },
      "APITagResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique tag ID"
          },
          "tag": {
            "type": "string",
            "description": "Tag string (lowercased, no spaces)"
          },
          "resource_type": {
            "type": "string",
            "enum": ["link", "quote"],
            "description": "Type of resource this tag is attached to"
          },
          "resource_id": {
            "type": "integer",
            "description": "ID of the resource this tag is attached to"
          },
          "created_by": {
            "type": "string",
            "description": "Username who created this tag"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the tag was created"
          }
        },
        "required": ["id", "tag", "resource_type", "resource_id", "created_at"],
        "example": {
          "id": 1,
          "tag": "golang",
          "resource_type": "link",
          "resource_id": 42,
          "created_by": "alice",
          "created_at": "2026-02-08T14:30:00Z"
        }
      },
      "APITagsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APITagResponse"
            }
          }
        },
        "required": ["data"]
      },
      "TagCreateRequest": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of tag strings (lowercased, no spaces, hyphens/underscores allowed)"
          },
          "user": {
            "type": "string",
            "description": "Username creating the tags (required for authorization if not using API key)"
          }
        },
        "required": ["tags"],
        "example": {
          "tags": ["golang", "web-dev", "tutorial"],
          "user": "alice"
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/links": {
      "get": {
        "summary": "List Links",
        "description": "Returns a paginated list of links. Can be filtered by user.",
        "tags": ["Links"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 50, max: 1000)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "user",
            "in": "query",
            "description": "Filter by submitter username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of links with pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILinksListResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation of links"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Link",
        "description": "Submits a new link. Links are always created even if the URL was previously submitted. Duplicate information is included in the response.",
        "tags": ["Links"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILinkCreateResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "123",
                  "description": "Link ID (plain text format)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (malformed JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid URL or missing fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/{id}": {
      "get": {
        "summary": "Get Link",
        "description": "Returns a single link by ID.",
        "tags": ["Links"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The link ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILinkResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Link",
        "description": "Deletes a link by ID. Requires API key authentication.",
        "tags": ["Links"],
        "security": [
          { "apiKey": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The link ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Link deleted successfully"
          },
          "401": {
            "description": "Unauthorized (missing API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (invalid API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/links/{id}/tags": {
      "get": {
        "summary": "List Tags for Link",
        "description": "Returns all tags associated with a link.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The link ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags for the link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Tags to Link",
        "description": "Adds tags to a link. Authorized if admin API key is present, or if within 10 minutes of link creation and username matches the original poster.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The link ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tags created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (malformed JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (not authorized to add tags)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid tag format)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace Tags on Link",
        "description": "Replaces all tags on a link. Same authorization rules as POST.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The link ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tags replaced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/link/": {
      "post": {
        "summary": "Submit a new Link (Deprecated)",
        "description": "DEPRECATED: Use POST /api/v1/links instead. Submits a new link using query parameters.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "The username of the submitter",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "The URL to submit",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Source of submission",
            "schema": {
              "type": "string",
              "enum": ["irc", "api", "web"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Link created"
          },
          "208": {
            "description": "Duplicate link created"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/link/{id}": {
      "get": {
        "summary": "Redirect to a Link (Deprecated)",
        "description": "DEPRECATED: Use GET /go/{id} instead. Redirects to the URL associated with the given ID.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the link",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to the target URL"
          },
          "404": {
            "description": "Link not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a link (Deprecated)",
        "description": "DEPRECATED: Use DELETE /api/v1/links/{id} instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the link",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}